Interfaces and classes to read datasets
Now time series datasets can be read using the classes instead of having to run a script saving to file a csv representing the training and one representing the testing. Thus, losing the information of the split used. Moreover, in such a way it is easier to use different datasets.