Refactor of reader package
The reader package has been refactored to simplify it and to simplify its meaning. The reader package now is solely devoted to data reading. It does not take care whether the type of data is for supervised or unsupervised learning. As well as the interfaces does not even assume the type of data: numerical or images