Yahoo reader and reader package refactoring
The yahoo reader has been completed and the reader package has been refactored such that it is simpler and more general. There is no distinction between univariate and multivariate time series reader classes. In fact, a reader for multivariate time series should have the same interface of a univariate given that a univariate time series is simply a multivariate time series with exactly one channel