Deep learning package refactoring
:white_check_mark: Tests passed

The package has been refactored such that the deep learning classes have a modular structure reducing the need of working directly with the classes of the library. The classes define just the shape they take as input, the preprocessing they do (sliding windows) and the helping utilities to manage input and output of neural network for prediction, scoring and classification. To develop a neural network doing prediction of the time series it is just necessary to write the model and a fitting function, then, they must be passed to the object of the class TSANNPredictor and the rest of the methods are implemented.